requirements.txt
openai==1.26.0
python-dotenv==1.0.1